feat: 메인페이지 로드맵 카운트 로직 수정 및 제목 페이지 연동

+++ src/main/java/com/example/hiddenpiece/controller/roadmap/RoadmapController.java
@@ -101,7 +101,7 @@ public class RoadmapController {
             @RequestParam(required = false) String date
     ) {
         if (date != null) {
-            return ResponseEntity.ok(roadmapService.countRoadmapsByCreatedAt());
+            return ResponseEntity.ok((int) roadmapService.countRoadmapsByCreatedAt());
         }
 
         return ResponseEntity.ok(roadmapService.countRoadmaps());
